+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Also want to check the development status, check the [commit history](https://gi
 
 ## [Released]
 ## [1.6.10] - 2021-xx-xx
+### Changed
+* Intellisense now works with dot notation (e.g. `object["key"]`)
+
 ### Fixed
 * [#104](https://github.com/zero-plusplus/vscode-autohotkey-debug/issues/104) Inherited members are not displayed in IntelliSense